# stormborne-booking-backend/docs/swagger.yaml
basePath: /
definitions:
...
  error.AppError:
    properties:
      code:
        type: string
      error: {}
      message:
        type: string
    type: object
...
  model.Slot:
    properties:
      endTime:
        type: string
      id:
        type: integer
      name:
        type: string
      startTime:
        type: string
    type: object
  request.AddScreenRequest:
    properties:
      name:
        type: string
    type: object
...
  request.ScheduleShowRequest:
    properties:
      cost:
        type: number
      date:
        type: string
      movieId:
        type: string
      rated:
        type: string
      screenId:
        type: integer
      slotId:
        type: integer
    required:
    - cost
    - date
    - movieId
    - screenId
    - slotId
    type: object
  request.UpdateAvatarRequest:
    properties:
      avatar_type:
        type: string
      avatar_url:
        type: string
      predefined_avatar_id:
        type: integer
    type: object
  response.AdminBookingViewItem:
    properties:
      id:
        type: integer
      movieName:
        type: string
      onlineCustomerId:
        type: integer
      showDate:
        type: string
      showId:
        type: integer
      startTime:
        type: string
      status:
        type: string
      totalPrice:
        type: number
      transactionId:
        type: string
    type: object
  response.AdminBookingViewResponse:
    properties:
      bookings:
        items:
          $ref: '#/definitions/response.AdminBookingViewItem'
        type: array
      total:
        type: integer
    type: object
  response.AvatarResponse:
    properties:
      avatar_type:
        type: string
      avatar_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  response.PredefinedAvatarListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/response.PredefinedAvatarResponse'
        type: array
      total:
        type: integer
    type: object
  response.PredefinedAvatarResponse:
    properties:
      gender:
        type: string
      id:
        type: integer
      url:
        type: string
    type: object
...
  response.ScheduleItem:
    properties:
      cost:
        type: number
      date:
        type: string
      id:
        type: integer
      movie:
        $ref: '#/definitions/response.ScheduledMovieBrief'
      screenId:
        type: integer
      slotId:
        type: integer
    type: object
  response.ScheduleShowResponse:
    properties:
      cost:
        type: number
      date:
        type: string
      id:
        type: integer
      movieId:
        type: string
      rated:
        type: string
      screenId:
        type: integer
      slotId:
        type: integer
    type: object
  response.ScheduledMovieBrief:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  response.ScreenOption:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  response.WeeklyScheduleResponse:
    properties:
      schedule:
        items:
          $ref: '#/definitions/response.ScheduleItem'
        type: array
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a skyfox
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Booking API
  version: "1.0"
paths:
...
  /admin/schedule/week:
    get:
      consumes:
      - application/json
      description: get weekly schedule by start date
      parameters:
      - description: Enter basic auth
        in: header
        name: Authorization
        required: true
        type: string
      - description: week start date in YYYY-MM-DD
        in: query
        name: startDate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.WeeklyScheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BasicAuth: []
      summary: Weekly Admin Schedule
      tags:
      - Admin-Schedule
  /admin/screens:
    get:
      consumes:
      - application/json
      description: get all screens
      parameters:
      - description: Enter basic auth
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ScreenOption'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BasicAuth: []
      summary: Admin Screens
      tags:
      - Admin-Schedule
    post:
      consumes:
      - application/json
      description: create a new screen for scheduling
      parameters:
      - description: Enter basic auth
        in: header
        name: Authorization
        required: true
        type: string
      - description: Add screen request
        in: body
        name: request
        schema:
          $ref: '#/definitions/request.AddScreenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ScreenOption'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BasicAuth: []
      summary: Add Screen
      tags:
      - Admin-Schedule
  /admin/shows:
    post:
      consumes:
      - application/json
      description: schedule a movie to a screen and slot on a date
      parameters:
      - description: Enter basic auth
        in: header
        name: Authorization
        required: true
        type: string
      - description: Schedule show request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ScheduleShowRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ScheduleShowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BasicAuth: []
      summary: Schedule Show
      tags:
      - Admin-Schedule
  /admin/shows/{showId}:
    delete:
      consumes:
      - application/json
      description: delete a scheduled show by id
      parameters:
      - description: Enter basic auth
        in: header
        name: Authorization
        required: true
        type: string
      - description: Show ID
        in: path
        name: showId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BasicAuth: []
      summary: Delete Show
      tags:
      - Admin-Schedule
  /admin/slots:
    get:
      consumes:
      - application/json
      description: get all slots
      parameters:
      - description: Enter basic auth
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Slot'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BasicAuth: []
      summary: Admin Slots
      tags:
      - Admin-Schedule
  /admin/view-bookings:
    get:
      description: Get all bookings with optional filters for admin
      parameters:
      - description: Enter bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Filter by start date (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: Filter by end date (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      - description: Filter by movie name
        in: query
        name: movieName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AdminBookingViewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BearerAuth: []
      summary: Get All Bookings
      tags:
      - Admin
...
  /avatar:
    get:
      consumes:
      - application/json
      description: Retrieve the avatar of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AvatarResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BearerAuth: []
      summary: Get my avatar
      tags:
      - Avatar
    put:
      consumes:
      - application/json
      description: Update the avatar for the authenticated user
      parameters:
      - description: Update Avatar Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateAvatarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.AvatarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      security:
      - BearerAuth: []
      summary: Update my avatar
      tags:
      - Avatar
  /avatars/predefined:
    get:
      consumes:
      - application/json
      description: Get predefined avatars optionally filtered by gender
      parameters:
      - description: Gender filter
        in: query
        name: gender
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PredefinedAvatarListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.AppError'
      summary: List predefined avatars
      tags:
      - Avatar
...
swagger: "2.0"